Refactorisation du code
