+++ include/LightSystem/SpotLight.hpp
